+++ src/common/MessageType.h
@@ -16,4 +16,5 @@ enum class MessageType : uint8_t {
   next_key_info,
   inject_input,
   inject_output,
+  notify,
 };